requirements.txt
Flask==2.3.2
waitress==2.1.2
python-dotenv==1.0.0
markdown2==2.4.6
llama-index==0.5.0
huggingface-hub==0.16.2
fitz==0.0.1.dev2
requests==2.31.0
Pillow==9.5.0
pywebview
multiprocessing==2.6.2.1
smtplib==3.1.4
imaplib2==2.58